testy

+++ Library Manager/Library Manager/UserControlMojeKsiazki.cs
@@ -24,46 +24,52 @@ namespace Library_Manager
 
         private void btnOddajKsiazke_Click(object sender, EventArgs e)
         {
-            int book_id = int.Parse(listViewMojeKsiazki.SelectedItems[0].Text);
-
-            string query = "SELECT user_login FROM users WHERE user_id LIKE " + my_id.ToString();
-
-            string user_login = form1obj.sendQueryRetString(query);
-
-            query = "DELETE FROM books_" + user_login + " WHERE book_id LIKE " + book_id.ToString();
-
-            string trash_result = form1obj.sendQueryRetString(query);
-
-            query = "DELETE FROM books_borrowed WHERE book_id LIKE " + book_id.ToString() + " AND user_id LIKE " + my_id.ToString();
-
-            trash_result = form1obj.sendQueryRetString(query);
-
-            query = "SELECT book_amount_available FROM books WHERE book_id LIKE " + book_id.ToString();
-
-            string result = form1obj.sendQueryRetString(query);
-
-            int books_after_return = int.Parse(result) + 1;
-
-            query = "UPDATE books SET book_amount_available = "+ books_after_return.ToString() +" WHERE book_id LIKE " + book_id.ToString();
-
-            trash_result = form1obj.sendQueryRetString(query);
-
-            List<ListViewItem> list_updated = new List<ListViewItem>();
-
-            query = "SELECT * FROM books_" + user_login;
-
-            list_updated = form1obj.sendQueryRetUserBooks(query);
-
-            listViewMojeKsiazki.Items.Clear();
-
-            foreach(ListViewItem item in list_updated)
+            if (listViewMojeKsiazki.SelectedItems.Count <= 0)
             {
-                listViewMojeKsiazki.Items.Add(item);
+                MessageBox.Show("Wybierz ksiazke");   
             }
+            else
+            {
+                int book_id = int.Parse(listViewMojeKsiazki.SelectedItems[0].Text);
+
+                string query = "SELECT user_login FROM users WHERE user_id LIKE " + my_id.ToString();
+
+                string user_login = form1obj.sendQueryRetString(query);
+
+                query = "DELETE FROM books_" + user_login + " WHERE book_id LIKE " + book_id.ToString();
+
+                string trash_result = form1obj.sendQueryRetString(query);
+
+                query = "DELETE FROM books_borrowed WHERE book_id LIKE " + book_id.ToString() + " AND user_id LIKE " + my_id.ToString();
+
+                trash_result = form1obj.sendQueryRetString(query);
+
+                query = "SELECT book_amount_available FROM books WHERE book_id LIKE " + book_id.ToString();
+
+                string result = form1obj.sendQueryRetString(query);
+
+                int books_after_return = int.Parse(result) + 1;
+
+                query = "UPDATE books SET book_amount_available = " + books_after_return.ToString() + " WHERE book_id LIKE " + book_id.ToString();
+
+                trash_result = form1obj.sendQueryRetString(query);
+
+                List<ListViewItem> list_updated = new List<ListViewItem>();
+
+                query = "SELECT * FROM books_" + user_login;
+
+                list_updated = form1obj.sendQueryRetUserBooks(query);
+
+                listViewMojeKsiazki.Items.Clear();
+
+                foreach (ListViewItem item in list_updated)
+                {
+                    listViewMojeKsiazki.Items.Add(item);
+                }
 
 
-            MessageBox.Show("Książka oddana.");
-
+                MessageBox.Show("Książka oddana.");
+            }
             
         }
     }
